% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{html2R}
\alias{html2R}
\title{Launch the 'html2r' Shiny app}
\usage{
html2R(file, theme = "cobalt", fontSize = 16)
}
\arguments{
\item{file}{path to a HTML file; can be missing}

\item{theme, fontSize}{options passed to
\code{\link[shinyAce:aceEditor]{aceEditor}}}
}
\description{
Shiny app allowing to convert HTML code to R code.
}
\examples{
# launch the Shiny app without file ####
if(interactive()) html2R()

# launch the Shiny app with a file ####
if(interactive()){
  html2R(system.file("example.html", package = "html2R"))
}
}
